+++ src/main/java/com/example/SVC/repository/VersionRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.List;
 
@@ -13,11 +14,11 @@ import java.util.List;
 public interface VersionRepository extends JpaRepository<DocumentVersion, Long> {
 
 
-    Optional<DocumentVersion> findByVersionAndDocument(double version, Document document);
+    Optional<DocumentVersion> findByVersionAndDocument(BigDecimal version, Document document);
     List<DocumentVersion> findAllByDocumentId(Long documentId);
     Optional<DocumentVersion> findById(Long id);
     @Query("SELECT MAX(d.version) FROM DocumentVersion d WHERE d.document.id = :documentId")
-    Double findNewestVersion(@Param("documentId") Long documentId);
+    BigDecimal findNewestVersion(@Param("documentId") Long documentId);
    
 
 }
